% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createhilldata.R
\name{createhilldata}
\alias{createhilldata}
\title{Create Hill Data}
\usage{
createhilldata(x, y, cat)
}
\arguments{
\item{x}{a numerical vector of length 1}

\item{y}{a numerical vector of length 1}

\item{cat}{a character/factor vector of length 1}
}
\value{
a list of length 3 with x,y and colour data
}
\description{
Creates data for hill chart
}
\examples{
createhilldata(2,400, "1")
}
\keyword{data}
\keyword{hill}
\keyword{plot}
