% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillplot.R
\name{hillplot}
\alias{hillplot}
\title{Create Hill Plot}
\usage{
hillplot(x, y, col_plot, x_lim, y_lim, stage, climb_name, avg_p = NULL,
  max_p = NULL)
}
\arguments{
\item{x}{a numerical vector used to plot x axis}

\item{y}{a numerical vector used to plot y axis}

\item{col_plot}{a character/factor vector to colour plot}

\item{x_lim}{max value for x lim}

\item{y_lim}{max value for y lim}

\item{stage}{stage number}

\item{climb_name}{climb name}

\item{avg_p}{average gradient}

\item{max_p}{max gradient}
}
\value{
plot hill
}
\description{
Creates hill chart
}
\examples{
d <- createhilldata(2,400, "1")
hillplot(x = d$x,
y = d$y,
col_plot = d$plot_col,
x_lim = 10,
y_lim = 500,
stage = 1,
climb_name = "IPSUM",
avg_p = 5,
max_p = 14)
}
\keyword{hill}
\keyword{plot}
