% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameasabove.R
\name{sameasabove}
\alias{sameasabove}
\title{Same As Above}
\usage{
sameasabove(x)
}
\arguments{
\item{x}{a vector of length > 2}
}
\description{
Replaces empty rows by the value above
}
\examples{
sameasabove(c(2,1,3,"","",4,NA))
}
\keyword{replace}
