% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.scrape.R
\name{stage.scrape}
\alias{stage.scrape}
\title{Create Stages Data}
\usage{
stage.scrape(years.range = c(1946, 2019))
}
\arguments{
\item{years.range}{a numerical vector of length 2 with range of years}
}
\value{
a list of length seq from year range 1 to year range 2
}
\description{
Scrapes stage data from Wikipedia
}
\examples{
stage.scrape(years.range = c(2019, 2019))
}
\keyword{scrape}
\keyword{stages}
\keyword{wiki}
