% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\docType{data}
\name{tt}
\alias{tt}
\title{Time Trial Data}
\format{A data frame with 21 rows and 5 variables:
\describe{
  \item{rider}{rider's name, character}
  \item{race}{name of the race, character}
  \item{time}{time in minutes, numeric}
  \item{speed}{speed in kmh, numeric}
  \item{len}{lenght in km, numeric}
}}
\source{
\url{http://procyclingstats.com/}
}
\usage{
tt
}
\description{
This dataset contains tt data form years 2019 and 2018
}
\keyword{datasets}
